+++ nts.uk/uk.at/at.ctx/function/nts.uk.ctx.at.function.dom/src/main/java/nts/uk/ctx/at/function/dom/indexreconstruction/repository/IndexReorgTableRepository.java
@@ -34,7 +34,7 @@ public interface IndexReorgTableRepository {
 	 * @param categoryNo the category no
 	 * @param tablePhysName the table phys name
 	 */
-	void delete(BigDecimal categoryNo, String tablePhysName);
+	void delete(int categoryNo, String tablePhysName);
 	
 	/**
 	 * Find all by category ids.
